render page based on pool type
